Add C++ solution for Missing_Number

+++ Introductory_Problems/Missing_Number/Missing_Number.cpp
@@ -19,8 +19,8 @@ int main() {
         cin>>x;
         sum += x;
     }
-    int total = (n*(n+1)/2);
-    int ans = total - sum;
+    ll total = (n*(n+1)/2);
+    ll ans = total - sum;
     cout<<ans<<"\n";
